put impls in .c files
