+++ docker-compose.yml
@@ -1,0 +1,6 @@
+version: '2'
+services:
+  test:
+    build:
+      context: .
+      dockerfile: test.Dockerfile
