+++ test.Dockerfile
@@ -1,0 +1,6 @@
+FROM swift:4.2
+COPY Sources/ Sources/
+COPY Tests/ Tests/
+COPY Package.swift Package.swift
+RUN swift build
+ENTRYPOINT swift test
